+++ packages/core/documentation/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@1.4.19...@kong-ui-public/documentation@1.4.20) (2025-12-08)
+
+**Note:** Version bump only for package @kong-ui-public/documentation
+
+
+
+
+
 ## [1.4.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/documentation@1.4.18...@kong-ui-public/documentation@1.4.19) (2025-12-07)