+++ packages/entities/entities-gateway-services/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.13.23](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.13.22...@kong-ui-public/entities-gateway-services@3.13.23) (2025-12-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-gateway-services
+
+
+
+
+
 ## [3.13.22](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-gateway-services@3.13.21...@kong-ui-public/entities-gateway-services@3.13.22) (2025-12-07)