+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.28](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.27...@kong-ui-public/entities-routes@3.16.28) (2025-12-08)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
 ## [3.16.27](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.16.26...@kong-ui-public/entities-routes@3.16.27) (2025-12-07)